+++ tools/conventions/declared_errors_test.go
@@ -1,4 +1,4 @@
-package errorregistry
+package conventions
 
 import (
 	"io/fs"